fix(golden): cast __row_id__ to Int64 before negating in the tie-break key (#870)
CI (datafusion spine parity + bench spill) hit
`InvalidOperationError: 'neg' operation not supported for dtype 'u32'`: the
real pipeline mints __row_id__ via with_row_index, which is UInt32, and the
most_complete composite sort key struct (len, -__row_id__) can't negate an
unsigned int. Local fixtures built __row_id__ from a Python list (Int64), so
they never reproduced it. Cast to Int64 before negating; the pick is unchanged
(lowest __row_id__ on a length tie). Verified on a real u32 with_row_index
frame + the spine-parity suite.

+++ packages/python/goldenmatch/goldenmatch/core/golden.py
@@ -396,7 +396,7 @@ def _stable_value_expr(
         # ``(len, -__row_id__)`` sorted descending == (len desc, __row_id__ asc),
         # and is a single filtered expression aligned with ``nn``.
         key = (
-            pl.struct([pl.col(len_alias), -pl.col("__row_id__")]).filter(mask)
+            pl.struct([pl.col(len_alias), -pl.col("__row_id__").cast(pl.Int64)]).filter(mask)
             if has_row_id
             else pl.col(len_alias).filter(mask)
         )
@@ -423,7 +423,7 @@ def _stable_rid_expr(
         # wins both (row-aligned provenance). See that helper for why one struct
         # key, not a list of filtered keys.
         key = (
-            pl.struct([pl.col(len_alias), -pl.col("__row_id__")]).filter(mask)
+            pl.struct([pl.col(len_alias), -pl.col("__row_id__").cast(pl.Int64)]).filter(mask)
             if has_row_id
             else pl.col(len_alias).filter(mask)
         )
